webapp/requirements.txt
Django==1.10.6
psycopg2==2.7.1
pycrypto==2.6.1
wheel==0.24.0
